+++ R/get_predictions_glmmTMB.R
@@ -7,6 +7,7 @@ get_predictions_glmmTMB <- function(model,
                                     value_adjustment,
                                     condition,
                                     interval = NULL,
+                                    bias_correction = FALSE,
                                     verbose = TRUE,
                                     ...) {
   # does user want standard errors?
@@ -52,7 +53,7 @@ get_predictions_glmmTMB <- function(model,
   }
 
   # check additional arguments, like number of simulations
-  additional_dot_args <- match.call(expand.dots = FALSE)[["..."]]
+  additional_dot_args <- list(...)
 
   if ("nsim" %in% names(additional_dot_args)) {
     nsim <- eval(additional_dot_args[["nsim"]])
@@ -64,15 +65,32 @@ get_predictions_glmmTMB <- function(model,
 
   if (type %in% c("zero_inflated", "zero_inflated_random")) {
 
-    prdat <- as.vector(stats::predict(
+    # prepare argument list
+    pr_args <- list(
       model,
       newdata = data_grid,
       type = "response",
       se.fit = FALSE,
       re.form = ref,
-      allow.new.levels = TRUE,
-      ...
-    ))
+      allow.new.levels = TRUE
+    )
+    # add `...` arguments
+    pr_args <- c(pr_args, additional_dot_args)
+    # remove arguments not supported by `predict()`
+    pr_args$sigma <- NULL
+
+    # call predict with argument list
+    prdat <- as.vector(do.call(stats::predict, pr_args))
+
+    # link function, for later use
+    lf <- insight::link_function(model)
+
+    # if we have bias-correction, we must also adjust the predictions
+    # this has not been done before, since we return predictions on
+    # the response scale directly, without any adjustment
+    if (isTRUE(bias_correction)) {
+      prdat <- linv(lf(prdat))
+    }
 
     if (se) {
 
@@ -141,7 +159,6 @@ get_predictions_glmmTMB <- function(model,
           # get link-function and back-transform fitted values
           # to original scale, so we compute proper CI
           if (!is.null(revar)) {
-            lf <- insight::link_function(model)
             predicted_data$conf.low <- exp(lf(predicted_data$conf.low) - tcrit * sqrt(revar))
             predicted_data$conf.high <- exp(lf(predicted_data$conf.high) + tcrit * sqrt(revar))
             predicted_data$std.error <- sqrt(predicted_data$std.error^2 + revar)
@@ -174,29 +191,29 @@ get_predictions_glmmTMB <- function(model,
     }
 
     # predictions conditioned on count or zi-component only
-
     if (type == "zi_prob") {
-      prdat <- stats::predict(
-        model,
-        newdata = data_grid,
-        type = "zlink",
-        se.fit = se,
-        re.form = ref,
-        allow.new.levels = TRUE,
-        ...
-      )
+      ptype <- "zlink"
       linv <- stats::plogis
     } else {
-      prdat <- stats::predict(
-        model,
-        newdata = data_grid,
-        type = "link",
-        se.fit = se,
-        re.form = ref,
-        allow.new.levels = TRUE,
-        ...
-      )
+      ptype <- "link"
     }
+
+    # prepare argument list
+    pr_args <- list(
+      model,
+      newdata = data_grid,
+      type = ptype,
+      se.fit = se,
+      re.form = ref,
+      allow.new.levels = TRUE
+    )
+    # add `...` arguments
+    pr_args <- c(pr_args, additional_dot_args)
+    # remove arguments not supported by `predict()`
+    pr_args$sigma <- NULL
+
+    # call predict with argument list
+    prdat <- do.call(stats::predict, pr_args)
 
     # did user request standard errors? if yes, compute CI
     if (se) {