+++ R/utils_set_attr.R
@@ -24,6 +24,7 @@
                                       margin = NULL,
                                       latent = FALSE,
                                       latent_thresholds = NULL,
+                                      bias_correction = FALSE,
                                       verbose = TRUE) {
   # check correct labels
   if (!is.null(x.axis.labels) && length(x.axis.labels) != length(stats::na.omit(unique(data$x)))) {
@@ -82,7 +83,7 @@
     }
     attr(data, "is.trial") <- ifelse(model_info$is_trial && inherits(model, "brmsfit"), "1", "0")
   }
-  attr(data, "link_inverse") <- insight::link_inverse(model)
+  attr(data, "link_inverse") <- .link_inverse(model, bias_correction = bias_correction)
   attr(data, "link_function") <- insight::link_function(model)
   attr(data, "n.trials") <- n.trials
   attr(data, "latent_thresholds") <- latent_thresholds